Image_Caption/requirements.txt
fire==0.4.0
loguru==0.6.0
matplotlib==3.5.1
nltk==3.7
numpy==1.21.6
pandas==1.3.5
Pillow==9.1.1
pycocoevalcap==1.2
PyYAML==6.0
skimage==0.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
